pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "workstation"
version = "0.1.0"
description = "Development Workstation Tools"
requires-python = ">=3.10"
authors = [
    {name = "Mark Carpenter", email = "mark@localhousebuyers.net"},
]
dependencies = [
    # Core dependencies
    "loguru==0.6.0",
    "sqlalchemy>=2.0.23",
    "watchdog>=3.0.0",
    "rich>=13.7.0",
    "click>=8.1.7",
    "psutil>=5.9.0",
    
    # AI and ML
    "crewai>=0.80.0",
    "langchain>=0.3.8",
    "langchain-openai>=0.2.9",
    "langchain-community>=0.3.8",
    "langchain-core>=0.1.26",
    "openai>=1.0.0",
    "duckduckgo-search",
    
    # Web and API
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "psycopg2-binary>=2.9.9",
    "jinja2>=3.0.0",
    "markdown>=3.3.0",
    "python-multipart>=0.0.5",
    "aiofiles>=0.8.0",
]

[project.scripts]
# Timeline commands
timeline = "timeline.cli:main"
dev-docs = "crews.visualization.dev_docs_server:main"
start-all = "scripts.start_services:start_services"

# Crew commands
run-crews = "crews.run_all_crews:main"
run-viz = "crews.visualization.server:main"

[tool.setuptools]
packages = [
    "crews",
    "timeline",
    "api_tester",
    "scripts"
]

[tool.black]
line-length = 88
target-version = ["py310"]
  